pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.5",
    "arabic-reshaper>=3.0.1",
    "axios>=0.4.0",
    "beautifulsoup4>=4.14.3",
    "browser-cookie3>=0.20.1",
    "cloudscraper>=1.2.71",
    "curl-cffi>=0.15.0",
    "ddgs>=9.14.1",
    "deep-translator>=1.11.4",
    "duckduckgo-search>=8.1.1",
    "email-validator>=2.3.0",
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "fpdf>=1.7.2",
    "fpdf2>=2.8.7",
    "gemini-webapi>=2.0.0",
    "gtts>=2.5.0",
    "gunicorn>=25.3.0",
    "httpx>=0.28.1",
    "lxml>=6.1.0",
    "lxml-html-clean>=0.4.4",
    "mega-py>=1.0.8",
    "pillow>=11.0.0",
    "psycopg2-binary>=2.9.12",
    "pycryptodomex>=3.20.0",
    "python-bidi>=0.6.7",
    "qrcode[pil]>=8.0",
    "readability-lxml>=0.8.4.1",
    "requests>=2.33.1",
    "trafilatura>=2.0.0",
    "youtube-transcript-api>=0.6.2",
]
